fully working architecture
This program takes a CSV file as input and creates an SSTable.
Generated SSTable files are zipped and uploaded to Amazon S3.
Zipped files are downloaded to the local machine
A JMX MBean instance is used to upload the SSTables into a running
instance of Cassandra.

Work to be done:
1. SStableloader creates input stream error due to “low transfer bytes”
when using the command sstableloader -d localhost userdata/users.
2. Automation scripts are being written in ruby and shell as well. Both
are in quality consistency checks and will be uploaded effectively
soon. This adds the ability to download dependencies to execute the
service.
3. Ruby scripts to perform operations on Amazon S3 is being written to
automate the process based on need. Will be completed after step 2.
4. Most of the code is open source and some have my own
implementations. So will port to open source based on need.
5. Code cleanup will be performed after the service is shippable as a
product.